refactored polling to pollingservice to better understand the class behaviour
