+++ build.gradle
@@ -1,12 +1,3 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.0.5.RELEASE")
-    }
-}
-
 plugins {
     id 'java'
     id 'application'
@@ -36,8 +27,9 @@ dependencies {
     testCompile group: 'junit', name: 'junit', version: '4.12'
     testCompile "org.mockito:mockito-core:2.+"
 
+    testCompile "org.mockito:mockito-core:1.+"
     testCompile group: 'junit', name: 'junit', version: '4.12'
-    testCompile('org.springframework.boot:spring-boot-start-test')
+    testCompile 'org.springframework.boot:spring-boot-start-test'
 }
 
 mainClassName = 'configuration.Application'
